Update sprite bomb

+++ include/monster.h
@@ -18,7 +18,7 @@ void monster_init(struct map* map, int x, int y, m_type type, int size, int life
 int monster_get_currentway(struct monster* monster);
 void monster_set_currentway(struct monster* monster, way_t dir);
 int monster_get_movetimer(struct monster* monster);
-void monster_set_movetimer(struct monster* monster, float timer);
+void monster_set_movetimer(struct monster* monster, int timer);
 int monster_get_aggr(struct monster* monster);
 
 int monster_pathfinding(struct map* map, struct player* player, struct list* mList, int* getDist);
@@ -31,7 +31,7 @@ int monster_get_nb_life(struct monster* monster);
 void monster_set_nb_life(struct monster* monster, int life);
 int monster_get_invincibility(struct monster* monster);
 void monster_set_invincibility(struct monster* monster, int invincibility);
-float monster_get_life_timer(struct monster* monster);
+int monster_get_life_timer(struct monster* monster);
 void monster_set_life_timer(struct monster* monster, int lifeTimer);
 
 struct list* monster_dec_nb_life(struct list* mList, int x, int y, struct game* game);
